Added file details
Strings are now sorted in the file in the exact same order as they are in the original (mixing bones and animations)
Missing bones are now properly added back in with 0 animations on them.
Values for animations are now rounded to the nearest 2e-06 to ease de-duplication (seems to be what the original tool did, results are within reason of the intent)
Padding was added at the end of channels to ensure multiple of 4.